+++ sport_companion/module-api/src/test/java/com/service/sport_companion/api/component/crawl/CrawlFixturesTest.java
@@ -4,9 +4,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
 
-import com.service.sport_companion.api.component.ClubsHandler;
-import com.service.sport_companion.api.component.FixtureHandler;
-import com.service.sport_companion.api.component.SeasonHandler;
+import com.service.sport_companion.api.component.club.ClubsHandler;
+import com.service.sport_companion.api.component.club.FixtureHandler;
+import com.service.sport_companion.api.component.club.SeasonHandler;
 import com.service.sport_companion.domain.entity.ClubsEntity;
 import com.service.sport_companion.domain.entity.FixturesEntity;
 import com.service.sport_companion.domain.entity.SeasonsEntity;
