+++ sport_companion/module-api/src/test/java/com/service/sport_companion/api/service/impl/AuthServiceImplTest.java
@@ -9,9 +9,9 @@ import static org.mockito.Mockito.when;
 
 import com.service.sport_companion.api.auth.jwt.JwtUtil;
 import com.service.sport_companion.api.auth.oauth.KakaoAuthHandler;
-import com.service.sport_companion.api.component.ClubsHandler;
+import com.service.sport_companion.api.component.club.ClubsHandler;
 import com.service.sport_companion.api.component.RedisHandler;
-import com.service.sport_companion.api.component.SupportedClubsHandler;
+import com.service.sport_companion.api.component.club.SupportedClubsHandler;
 import com.service.sport_companion.api.component.UserHandler;
 import com.service.sport_companion.core.exception.GlobalException;
 import com.service.sport_companion.domain.entity.ClubsEntity;
